backend/netnotes/notes/serializers.py
from rest_framework import serializers
from .models import Note


class NoteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Note
        fields = ["id", "slug", "title", "description"]
        read_only_fields = ("slug",)
