backend/pyproject.toml
[tool.poetry]
name = "netnotes-api"
version = "0.1.0"
description = ""
authors = ["esskayesss <imesskayesss@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
uuid = "^1.30"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
